Integrate the checked grid cardinality and direct Apollo axis-storage changes with the provider-first GPU and locked-CI migration. Keep package-scoped locked commands and merge both PM surfaces.

Refs: backlog.md#KW-CI-052

+++ crates/kwavers/examples/pinn_real_time_inference.rs
@@ -181,7 +181,7 @@ fn run_performance_benchmark() -> KwaversResult<()> {
         let _result = simulate_inference(&test_input);
 
         let elapsed = start.elapsed().as_micros();
-        total_time += elapsed as u128;
+        total_time += elapsed;
     }
 
     let avg_latency = total_time as f64 / num_samples as f64;
